requirements.txt
streamlit==1.49.1
langchain==0.3.27
langchain-google-genai==2.1.10
langchain-core==0.3.75
langchain-community==0.3.29
langchain-text-splitters==0.3.10
python-docx==1.2.0
PyPDF2==3.0.1
python-dotenv==1.1.1
google-ai-generativelanguage==0.6.18
transformers>=4.56.0
torch>=2.8.0
accelerate>=1.10.1
